requirements.txt
amqp==5.0.9
anyio==3.5.0
asgiref==3.4.1
async-generator==1.10
attrs==21.4.0
beautifulsoup4==4.10.0
billiard==3.6.4.0
black==21.12b0
celery==5.2.3
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
colorlog==4.8.0
commonmark==0.9.1
cryptography==36.0.1
fastapi==0.73.0
flake8==4.0.1
flake8-polyfill==1.0.2
future==0.18.2
gitdb==4.0.9
GitPython==3.1.26
h11==0.13.0
idna==3.3
importlib-resources==5.4.0
ipython-genutils==0.2.0
jsonschema==4.4.0
jupyter-core==4.9.1
kombu==5.2.3
mando==0.6.4
mccabe==0.6.1
motor==2.5.1
mypy-extensions==0.4.3
nbformat==5.1.3
outcome==1.1.0
pathspec==0.9.0
platformdirs==2.4.1
plotly==4.14.3
progress==1.4
prompt-toolkit==3.0.24
pycodestyle==2.8.0
pycparser==2.21
pydantic==1.9.0
pyflakes==2.4.0
Pygments==2.11.2
pymongo==3.12.3
pyOpenSSL==22.0.0
pyrsistent==0.18.1
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
radon==4.0.0
requests==2.27.1
retrying==1.3.3
rich==9.13.0
selenium==4.1.0
six==1.16.0
smmap==5.0.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.1
starlette==0.17.1
tabulate==0.8.9
tomli==1.2.3
traitlets==5.1.1
trio==0.19.0
trio-websocket==0.9.2
typing-extensions==3.10.0.2
urllib3==1.26.8
uvicorn==0.17.0
vine==5.0.0
wcwidth==0.2.5
wsproto==1.0.0
zipp==3.7.0
